added styling for the sidebar
