+++ frontend/src/modules/layout/sidebar.css
@@ -1,0 +1,9 @@
+section.sidebar {
+  position: absolute;
+  height: 100vh;
+  width: 20%;
+  z-index: 1001;
+  color: rgba(240, 240, 240, 1);
+  background-color: rgba(33, 33, 33, 0.8);
+  top: 132px;
+}
