+++ frontend/src/modules/map/route-details.css
@@ -15,8 +15,8 @@ button {
   border: none;
   font-weight: bolder;
   width: 100px;
-
 }
+
 label {
 
 }
